Add Ev3OutputPort.PORT_D and update tests
