+++ src/main/java/com/pileproject/drivecommand/model/ev3/port/Ev3InputPort.java
@@ -24,6 +24,6 @@ public class Ev3InputPort extends InputPort {
 
     @Override
     public String toString() {
-        return "[EV3] INPUT-PORT-" + mPort;
+        return "[EV3] INPUT-PORT-" + (mPort + 1);
     }
 }
