+++ src/main/java/com/pileproject/drivecommand/model/nxt/port/NxtOutputPort.java
@@ -23,9 +23,7 @@ public class NxtOutputPort extends OutputPort {
 
 	@Override
 	public String toString() {
-		final int A = 65;
-
-		return "[NXT] OUTPUT-PORT-" +  (char)(A + mPort);
+		return "[NXT] OUTPUT-PORT-" +  (char)('A' + mPort);
 	}
 }
